Crate rust_wasm_router

rust_wasm_router

version: 0.4.3 date: 2020-04-25 authors: Luciano Bestia
wasm router for local hash routes

local router with hash

I needed a router for local hash routes in rust wasm.
This library contains the generic parts.

In the project add a file/mod with code like this example:
https://github.com/LucianoBestia/mem6_game/blob/master/mem6/src/routerimplmod.rs

Modules

routermod
routermod - A simple #-fragment router for dodrio html templating This is the generic module. All the specifics for a website are isolated in the module routerimplmod.